refactor: Counter 관련 로직을 useCounter 커스텀 훅으로 분리
