+++ src/components/Counter/Counter.tsx
@@ -1,6 +1,7 @@
-import { ChangeEventHandler } from 'react';
 import { css, styled } from 'styled-components';
 import { MinusIcon, PlusIcon } from '../../assets/icons';
+import useCounter from './useCounter';
+import { ChangeEventHandler } from 'react';
 
 export type CounterSize = 'medium' | 'small';
 
@@ -11,29 +12,13 @@ interface CounterProps {
 }
 
 const Counter = ({ count, onChangeCount, size = 'medium' }: CounterProps) => {
-  const increaseCount = () => {
-    if (count >= 99) return;
+  const { increaseCount, decreaseCount, handleCountChange } = useCounter(
+    count,
+    onChangeCount,
+  );
 
-    onChangeCount(count + 1);
-  };
-
-  const decreaseCount = () => {
-    if (count <= 0) return;
-
-    onChangeCount(count - 1);
-  };
-
-  const handleCountChange: ChangeEventHandler<HTMLInputElement> = (e) => {
-    const value = e.target.value;
-
-    if (!isNumber(value)) return;
-
-    onChangeCount(Number(value));
-  };
-
-  const isNumber = (value: string) => {
-    const regex = /^\d*$/;
-    return regex.test(value);
+  const handleChange: ChangeEventHandler<HTMLInputElement> = (e) => {
+    handleCountChange(e.target.value);
   };
 
   return (
@@ -41,12 +26,7 @@ const Counter = ({ count, onChangeCount, size = 'medium' }: CounterProps) => {
       <CounterButton onClick={decreaseCount}>
         <MinusIcon size={size} />
       </CounterButton>
-      <Input
-        type="text"
-        maxLength={2}
-        value={count}
-        onChange={handleCountChange}
-      />
+      <Input type="text" maxLength={2} value={count} onChange={handleChange} />
       <CounterButton onClick={increaseCount}>
         <PlusIcon size={size} />
       </CounterButton>
